feat: workflows paths (#121)
Workflow jobs increase or decrease the workflow-job active paths. The number of active paths depends on the number of simultaneous jobs being executed within a workflow execution.
The number of active paths will reach 0 when the last node of every path is executed.

when a job failed to start that path won't reach 0. In this case the workflow will never trigger the finished event. But this is a special case


* fix: notification tasks is handled by the event handlers

* update paths

* code comment

* comment